packages/agent-core/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Shared library for the agent platform"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0rc5",
    "agent-framework-azure-ai>=1.0.0rc5",
    "agent-framework-a2a>=1.0.0b260319",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "pyyaml>=6.0",
]

[project.scripts]
agent-scaffold = "agent_core.scaffold:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_core"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
